Crate diem_json_rpc

JSON RPC endpoint

Used as public API interface for interacting with Full Nodes It serves HTTP API requests from various external clients (such as wallets)

Protocol specification: https://www.jsonrpc.org/specification

Module organization: ├── methods.rs # contains all available JSON RPC method handlers ├── runtime.rs # implementation of JSON RPC protocol over HTTP ├── tests.rs # tests

Re-exports

pub use runtime::bootstrap;

pub use runtime::bootstrap_from_config;

Modules

data

errors

fuzzing
Fuzzer for JSON RPC service

response

runtime

stream_rpc
JSON RPC endpoint

util

views

Macros

gen_request_params

log_response

Functions

test_bootstrap
Creates JSON RPC server for a Validator node Should only be used for unit-tests