Function diem_json_rpc::stream_rpc::startup::get_stream_routes

pub fn get_stream_routes(
    config: &StreamConfig, 
    content_length_limit: u64, 
    diem_db: Arc<dyn DbReader>
) -> BoxedFilter<(impl Reply,)>

Gets all routes for all streaming endpoints Each transport is responsible for handling it’s own versioning and endpoints This function also handles disabling/enabling streaming- returning a 404 if not enabled